Unit testing

+++ src/TrackedModel.php
@@ -1,16 +1,13 @@
 <?php
 
 
-namespace Cbwar\Laravel\ModelChanges\Models;
+namespace Cbwar\Laravel\ModelChanges;
 
-use Cbwar\Laravel\ModelChanges\Errors\TrackableError;
-use Cbwar\Laravel\ModelChanges\Observers\TrackedModelObserver;
-use Cbwar\Laravel\ModelChanges\Traits\TrackableFields;
 use Illuminate\Database\Eloquent\Model;
 
 abstract class TrackedModel extends Model
 {
-    use TrackableFields;
+    use TrackableFieldsTrait;
 
     /**
      * Registered events
@@ -40,7 +37,9 @@ abstract class TrackedModel extends Model
 
         // Remove tracks
         static::deleting(function (TrackedModel $model) {
-            $model->tracks()->delete();
+            if (!method_exists($model, 'isForceDeleting') || true === $model->isForceDeleting()) {
+                $model->tracks()->delete();
+            }
         });
 
         static::$tracking_sentences_default = [
